refactor(interop): move log helpers into op-core/interop/messages
Moves MessageFromLog, DecodeExecutingMessageLog, and LogToLogHash out of
op-supervisor/supervisor/backend/processors and into
op-core/interop/messages. These functions are pure log-decoding helpers
that operate entirely on op-core/interop/messages types; they have no
coupling to supervisor backend state. Several non-supervisor consumers
(op-interop-filter, op-interop-mon, op-supernode, op-program) were
reaching into the supervisor processors package just for these helpers.

After this change there are no external imports of
op-supervisor/supervisor/backend/processors. The processors package
remains as the supervisor's internal log-ingestion pipeline
(chain_processor, log_processor, client) and consumes the helpers via
op-core/interop/messages alongside the rest of its callers.

EventDecoderFn is consolidated into log_processor.go since it is only
used there.
