+++ op-supervisor/supervisor/backend/processors/log_processor.go
@@ -6,11 +6,12 @@ import (
 
 	"github.com/ethereum/go-ethereum/common"
 	ethTypes "github.com/ethereum/go-ethereum/core/types"
-	"github.com/ethereum/go-ethereum/crypto"
 
 	"github.com/ethereum-optimism/optimism/op-core/interop/messages"
 	"github.com/ethereum-optimism/optimism/op-service/eth"
 )
+
+type EventDecoderFn func(*ethTypes.Log) (*messages.ExecutingMessage, error)
 
 type LogStorage interface {
 	SealBlock(chain eth.ChainID, block eth.BlockRef) error
@@ -27,7 +28,7 @@ func NewLogProcessor(chain eth.ChainID, logStore LogStorage) LogProcessor {
 	return &logProcessor{
 		chain:        chain,
 		logStore:     logStore,
-		eventDecoder: DecodeExecutingMessageLog,
+		eventDecoder: messages.DecodeExecutingMessageLog,
 	}
 }
 
@@ -37,7 +38,7 @@ func (p *logProcessor) ProcessLogs(_ context.Context, block eth.BlockRef, rcpts 
 	for _, rcpt := range rcpts {
 		for _, l := range rcpt.Logs {
 			// log hash represents the hash of *this* log as a potentially initiating message
-			logHash := LogToLogHash(l)
+			logHash := messages.LogToLogHash(l)
 			// The log may be an executing message emitted by the CrossL2Inbox
 			execMsg, err := p.eventDecoder(l)
 			if err != nil {
@@ -54,14 +55,4 @@ func (p *logProcessor) ProcessLogs(_ context.Context, block eth.BlockRef, rcpts 
 		return fmt.Errorf("failed to seal block %s: %w", block.ID(), err)
 	}
 	return nil
-}
-
-// LogToLogHash transforms a log into a hash that represents the log.
-// it is the concatenation of the log's address and the hash of the log's payload,
-// which is then hashed again. This is the hash that is stored in the log storage.
-// The address is hashed into the payload hash to save space in the log storage,
-// and because they represent paired data.
-func LogToLogHash(l *ethTypes.Log) common.Hash {
-	payloadHash := crypto.Keccak256Hash(messages.LogToMessagePayload(l))
-	return messages.PayloadHashToLogHash(payloadHash, l.Address)
 }
